+++ VTK/Hybrid/vtkLegendBoxActor.cxx
@@ -405,7 +405,7 @@ int vtkLegendBoxActor::RenderOpaqueGeometry(vtkViewport *viewport)
       str = this->TextMapper[i]->GetInput();
       if ( str ) //if there is a string
         {
-        length = strlen(str);
+        length = static_cast<int>(strlen(str));
         if ( length > maxLength )
           {
           maxLength = length;
